Add js Game and Player initial
Add player factory function, with score as private variable that can be updated using public functions, and function that filters the gameboard through signs to its indexes. the rest of the game logic relies on those indexes.

Add game module with functions to check an array (which just got an additional cell) for rows that win a round of tic tac toe.