Create functional game rules
used async await function in order to create the player move. As a result needed to make some other function async, tried to fix it otherwise but didnt work. need to check how to do that in the future.

linked html with the game play function, want to somehow make it cleaner though.

forgot to commit a few times.