Add html and base css
Added game.html which is where the game itself is played, the index html is where the user decides how many rounds etc.
the styles css uses pseudoelements for the x's and o's, this is not best practice usually, will try to find another way to do this that's more accessible. Will have to find a font that fits what I'm looking for, can either enter the letter with js, or create two divs within each grid item one which will be displayed when clicked.