Add basic AImove
This ai basically checks which cells on the board haven't been taken yet and then chooses randomly one of them.
Fixed the playerXController in the main.js, the typo made it so one of the players didn't get a controller.
Just generally the addition of controller, and made the default controls apparent to the user in the index html.

TODO:
- add some setTimeout so that the human player can see the moves being made, otherwise it's a little too fast.