Add ai intermediate level and levels functionality
The primary ai was a beginner one, it looked for free cells and chose randomly which to fill.
The medium one is smarter, its first two moves are random but after that it checks first if it can win by filling one of the remaining cells, if it cant win it checks if the opponent can win and fills that cell, if neither it fills a cell at random.
This is still not the unbeatable ai, not sure if ill build it, but if so it will be easy to add in.

Completed the html, css and js for this addition - when choosing an AI player the option to choose the ai level is opened, if you choose human, it disappears again. Changed some css so it feeds off the same styling as previous styling.
Currently it works well.